requirements.txt
attrs==26.1.0
awslambdaric==4.0.0
blinker==1.9.0
click==8.3.1
dotenv==0.9.9
flasgger==0.9.7.1
Flask==3.1.3
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mistune==3.2.0
packaging==26.0
PyJWT==2.12.1
PyMySQL==1.1.2
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
serverless-wsgi==3.1.0
simplejson==3.20.2
six==1.17.0
snapshot-restore-py==1.0.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
Werkzeug==3.1.6
